+++ cmd/config.go
@@ -4,7 +4,6 @@ import (
 	"clx/file"
 	"clx/settings"
 	"fmt"
-	"os"
 
 	"github.com/spf13/cobra"
 )
@@ -14,7 +13,7 @@ func init() {
 }
 
 var createConfigCmd = &cobra.Command{
-	Use:   "create_config",
+	Use:   "config",
 	Short: "Create an example config file",
 	Long: "Create an example config file in ~/.config/circumflex/config.env.\n" +
 		"If a config file already exists, it will be overwritten.",
@@ -23,13 +22,9 @@ var createConfigCmd = &cobra.Command{
 	Run: func(cmd *cobra.Command, args []string) {
 		err := file.WriteToFileInConfigDir(file.PathToConfigFile(), settings.GetConfigFileContents())
 		if err != nil {
-			fmt.Println(err)
-
-			os.Exit(1)
+			panic(err)
 		}
 
 		fmt.Println("Example config file written to ~/.config/circumflex/config.env")
-
-		os.Exit(0)
 	},
 }