+++ CHANGELOG.md
@@ -6,6 +6,7 @@ _WIP_
 **Cosmetic**
 - Code snippets are now highlighted in magenta instead of blue
 - Mark as read setting now turned on by default
+- Rename create_config command to config
 
 ## 1.15
 _2021-07-22_